Adds event listener to form submission
Adds a new `FormSubmitObserver` class that adds a new listener to the `FormController` element for the `"submit"` event.

This change guarantees that the form will be validated when the user clicks on the submit input/button, or when they press enter when a form field is focused.

This change also eliminates the need to set `data-action="click->form#submit"` on the submit input/button.

The `submit` method of the `FormController` was removed and a new `validate` method was added. Users that want to trigger form validations programmatically may use the new method.